+++ terminus/const.py
@@ -1,0 +1,3 @@
+DEFAULT_PANEL = "Terminus"
+EXEC_PANEL = "Terminus Build Results"
+CONTINUATION = "\u200b\u200c\u200b"
