+++ terminus/mouse.py
@@ -5,7 +5,8 @@ import re
 import logging
 import webbrowser
 
-from .terminal import Terminal, CONTINUATION
+from .const import CONTINUATION
+from .terminal import Terminal
 from .utils import highlight_key
 
 logger = logging.getLogger('Terminus')
@@ -14,10 +15,10 @@ rex = re.compile(
     r'''(?x)
     \b(?:
         https?://(?:(?:[a-zA-Z0-9\-._]+)|localhost)|  # http://
-        www\.[a-zA-Z0-9\-_]+(?:\.[a-zA-Z0-9\-._]+)+                         # www.
+        www\.[a-zA-Z0-9\-_]+(?:\.[a-zA-Z0-9\-._]+)+   # www.
     )
-    /?[a-zA-Z0-9\-._?,!'(){}\[\]/+&@%$#=:"|~;]*                             # url path and query string
-    [a-zA-Z0-9\-_~:/#@$*+=]                                                 # allowed end chars
+    /?[a-zA-Z0-9\-._?,!'(){}\[\]/+&@%$#=:"|~;]*       # url path and query string
+    [a-zA-Z0-9\-_~:/#@$*+=]                           # allowed end chars
     ''')
 
 URL_POPUP = """
